Move gst_iterate from gst_hacks into gst_utils
It is not a hack, it uses the interfaces provided by py-gi.  I'm doing this
now as I'm about to add an additional (true) hack to gst_hacks.
