+++ _stbt/gst_hacks.py
@@ -130,19 +130,3 @@ def test_map_sample_without_buffer():
             assert False
     except ValueError:
         pass
-
-
-def gst_iterate(gst_iterator):
-    """Wrap a Gst.Iterator to expose the Python iteration protocol.  The
-    gst-python package exposes similar functionality on Gst.Iterator itself so
-    this code should be retired in the future once gst-python is broadly enough
-    available."""
-    result = Gst.IteratorResult.OK
-    while result == Gst.IteratorResult.OK:
-        result, value = gst_iterator.next()
-        if result == Gst.IteratorResult.OK:
-            yield value
-        elif result == Gst.IteratorResult.ERROR:
-            raise RuntimeError("Iteration Error")
-        elif result == Gst.IteratorResult.RESYNC:
-            raise RuntimeError("Iteration Resync")
